Move all implementations into the same files as abstract classes
