+++ api/src/infrastructure/tests/infrastructure/utilities/test_text_normalizer.py
